Start importing Kolibri
Add a kolibri submodule and build the wheel.
Then try to access the wsgi application object from the webworker,
as a way of checking we can load everything up to that point.

There are a lot of dependencies to handle. Many can be imported by pip.
Some need special treatment.

At this point, much loading is done but it fails importing sqlalchemy, a
binary module.